02-applications/youtube-digest-bot/pyproject.toml
[project]
name = "youtube-digest-bot"
version = "0.1.0"
description = "Automated YouTube video summarization bot with email digest"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.3.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "youtube-transcript-api>=0.6.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "C4", "DTZ", "T20", "RET", "SIM", "ARG", "PTH", "PL", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.black]
line-length = 88
target-version = ['py311']